# src/main/resources/openapi.yaml
openapi: 3.0.0
info:
  title: Home Budget Tracker API
  version: 1.0.0
  description: API documentation for Home Budget Tracker

paths:
  /api/v1/transaction:
    get:
      summary: Get all transactions by user
      description: Get all transactions by user with pagination, optionally filtered by currency code
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 10
        - name: code
          in: query
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Page'
        '401':
          description: Unauthorized

    post:
      summary: Create a new transaction
      description: Create a new transaction
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                amount:
                  type: number
                date:
                  type: string
                  format: date
                category:
                  type: string
                currency:
                  type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                  status:
                    type: string
                    example: Created
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                type: object
                properties:
                  businessErrorCode:
                    type: integer
                    example: 306
                  businessErrorDescription:
                    type: string
                    example: Missing or invalid request body element
                  validationErrors:
                    type: array
                    items:
                      type: string
        '401':
          description: Unauthorized

  /api/v1/transaction/category:
    get:
      summary: Get all transactions by user and category
      description: Get all transactions by user and category with pagination, optionally filtered by currency code
      parameters:
        - name: categoryName
          in: query
          required: true
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 10
        - name: code
          in: query
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Page'
        '401':
          description: Unauthorized
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                type: object
                properties:
                  businessErrorCode:
                    type: integer
                    example: 402
                  businessErrorDescription:
                    type: string
                    example: Record does not exist or is not accessible

  /api/v1/transaction/date:
    get:
      summary: Get all transactions by user and date between
      description: Get all transactions by user and date between with pagination, optionally filtered by currency code
      parameters:
        - name: startDate
          in: query
          required: true
          schema:
            type: string
            format: date
        - name: endDate
          in: query
          required: true
          schema:
            type: string
            format: date
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 10
        - name: code
          in: query
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Page'
        '401':
          description: Unauthorized
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                type: object
                properties:
                  businessErrorCode:
                    type: integer
                    example: 402
                  businessErrorDescription:
                    type: string
                    example: Record does not exist or is not accessible

  /api/v1/transaction/category-date:
    get:
      summary: Get all transactions by user, category, and date between
      description: Get all transactions by user, category, and date between with pagination, optionally filtered by currency code
      parameters:
        - name: categoryName
          in: query
          required: true
          schema:
            type: string
        - name: startDate
          in: query
          required: true
          schema:
            type: string
            format: date
        - name: endDate
          in: query
          required: true
          schema:
            type: string
            format: date
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 10
        - name: code
          in: query
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Page'
        '401':
          description: Unauthorized
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                type: object
                properties:
                  businessErrorCode:
                    type: integer
                    example: 402
                  businessErrorDescription:
                    type: string
                    example: Record does not exist or is not accessible

  /api/v1/transaction/{id}:
    delete:
      summary: Delete a transaction
      description: Delete a transaction
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: No Content
        '401':
          description: Unauthorized
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                type: object
                properties:
                  businessErrorCode:
                    type: integer
                    example: 402
                  businessErrorDescription:
                    type: string
                    example: Record does not exist or is not accessible

  /api/v1/transaction/sum-positive:
    get:
      summary: Sum positive amounts
      description: Sum positive amounts filtered by currency code
      parameters:
        - name: code
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  total:
                    type: number
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                type: object
                properties:
                  businessErrorCode:
                    type: integer
                    example: 400
                  businessErrorDescription:
                    type: string
                    example: "Missing request parameter: code"
        '401':
          description: Unauthorized

  /api/v1/transaction/sum-negative:
    get:
      summary: Sum negative amounts
      description: Sum negative amounts filtered by currency code
      parameters:
        - name: code
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  total:
                    type: number
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                type: object
                properties:
                  businessErrorCode:
                    type: integer
                    example: 400
                  businessErrorDescription:
                    type: string
                    example: "Missing request parameter: code"
        '401':
          description: Unauthorized

  /api/v1/transaction/sum:
    get:
      summary: Sum amounts
      description: Sum amounts filtered by currency code
      parameters:
        - name: code
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  total:
                    type: number
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                type: object
                properties:
                  businessErrorCode:
                    type: integer
                    example: 400
                  businessErrorDescription:
                    type: string
                    example: "Missing request parameter: code"
        '401':
          description: Unauthorized

  /api/v1/category:
    get:
      summary: Get all categories
      description: Get all categories
      parameters: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Category'
        '401':
          description: Unauthorized

    post:
      summary: Create a new category
      description: Create a new category
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                  status:
                    type: string
                    example: Created
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                type: object
                properties:
                  businessErrorCode:
                    type: integer
                    example: 306
                  businessErrorDescription:
                    type: string
                    example: Missing or invalid request body element
                  validationErrors:
                    type: array
                    items:
                      type: string
        '401':
          description: Unauthorized

  /api/v1/category/{id}:
    delete:
      summary: Delete a category
      description: Delete a category
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: No Content
        '401':
          description: Unauthorized
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                type: object
                properties:
                  businessErrorCode:
                    type: integer
                    example: 402
                  businessErrorDescription:
                    type: string
                    example: Record does not exist or is not accessible

components:
  schemas:
    Page:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/Transaction'
        totalPages:
          type: integer
        totalElements:
          type: integer
        size:
          type: integer
        number:
          type: integer
        first:
          type: boolean
        last:
          type: boolean

    Transaction:
      type: object
      properties:
        id:
          type: integer
        amount:
          type: number
        date:
          type: string
          format: date
        category:
          type: string
        currency:
          type: string

    Category:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string